\name{as.data.frame.nc_slice}
\alias{as.data.frame.nc_slice}
\title{Convert a "slice" extracted from a NetCDF file into a data.frame}
\usage{
\method{as.data.frame}{nc_slice}(x, ...)
}
\arguments{
  \item{x}{a NetCDF slice object, extracted via
  \code{\link{nc_slice}}}

  \item{...}{passed to generic method}
}
\description{
Convert a "slice" extracted from a NetCDF file into a
data.frame
}
\seealso{
\code{\link{nc_slice}}
}

