% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_slice.R
\name{as.data.frame.ncslice4}
\alias{as.data.frame.ncslice4}
\title{Convert a NetCDF slice into a data.frame}
\usage{
\method{as.data.frame}{ncslice4}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a NetCDF slice object, extracted via \code{\link{ncvar_slice}}.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and converting column names (to syntactic names: see \code{\link[base]{make.names}}) is optional.}

\item{...}{ignored.}
}
\description{
Convert a slice read from a NetCDF file with \code{\link{ncvar_slice}} into a data.frame for further manipulation of the data.
}
\examples{
nc <- nc_open(system.file("extdata", "sresa1b_ncar_ccsm3.nc", package="ncdf4helpers"))
x <- ncvar_slice(nc, "tas")
nc_close(nc)
head(as.data.frame(x))
}
\seealso{
\code{\link{ncvar_slice}} to read the slice and \code{\link{as.xyz.ncslice4}} to convert it into another format.
}

