% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_slice.R
\name{as.xyz.ncslice4}
\alias{as.xyz.ncslice4}
\title{Convert a NetCDF slice into xyz list}
\usage{
\method{as.xyz}{ncslice4}(x)
}
\arguments{
\item{x}{a NetCDF slice object, extracted via \code{\link{ncvar_slice}}}
}
\description{
Convert a slice read from a NetCDF file with \code{\link{ncvar_slice}} into a x,y,z list suitable for \code{\link[graphics]{image}}, \code{\link[graphics]{contour}}, \code{\link[graphics]{persp}}, etc.
}
\examples{
nc <- nc_open(system.file("extdata", "sresa1b_ncar_ccsm3.nc", package="ncdf4helpers"))
x <- ncvar_slice(nc, "tas")
nc_close(nc)
# inspect and transform the result
str(x)
x <- as.xyz(x)
str(x)
}
\seealso{
\code{\link{ncvar_slice}} to read the slice and \code{\link{as.data.frame.ncslice4}} to convert it into another format.
}

