% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_slice.R
\name{as.xyz}
\alias{as.xyz}
\title{Coerce to a x,y,z list}
\usage{
as.xyz(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Generic function to convert an object into an x,y,z list suitable for \code{\link[graphics]{image}}, \code{\link[graphics]{contour}}, \code{\link[graphics]{persp}}, etc.
}

