% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\docType{data}
\name{med}
\alias{med}
\title{Altitude over the Mediterranean Sea}
\format{a \code{data.frame} with columns: \code{lon}, \code{lat} and \code{altitude}}
\source{
http://maps.ngdc.noaa.gov/viewers/wcs-client/
}
\usage{
data(med)
}
\description{
Altitude data from NGDC interpolated over a coarse grid
}

