\name{plot.nc_slice}
\alias{plot.nc_slice}
\title{Plot a "slice" extracted from a NetCDF file}
\usage{
\method{plot}{nc_slice}(x, ...)
}
\arguments{
  \item{x}{a NetCDF slice object, extracted via
  \code{\link{nc_slice}}}

  \item{...}{passed to \code{image}}
}
\description{
Plot a "slice" extracted from a NetCDF file
}
\seealso{
\code{\link{nc_slice}}
}

