% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_slice.R
\name{image.ncslice4}
\alias{contour.ncslice4}
\alias{image.ncslice4}
\alias{persp.ncslice4}
\alias{plot.ncslice4}
\title{Plot a slice extracted from a NetCDF file}
\usage{
\method{image}{ncslice4}(x, ...)

\method{plot}{ncslice4}(x, ...)

\method{contour}{ncslice4}(x, ...)

\method{persp}{ncslice4}(x, ...)
}
\arguments{
\item{x}{a NetCDF slice object, extracted via \code{\link{ncvar_slice}}}

\item{...}{passed to the plotting function}
}
\description{
Plot a slice extracted from a NetCDF file
}
\examples{
nc <- nc_open(system.file("extdata", "sresa1b_ncar_ccsm3.nc", package="ncdf4helpers"))
x <- ncvar_slice(nc, "tas", lon=10:30, lat=30:50)
nc_close(nc)
image(x)
contour(x, add=TRUE)
persp(x, theta=210, phi=30)

}
\seealso{
\code{\link{ncvar_slice}}
}

